initial reshuffle
