+++ src/net/jnxnetwork.h
@@ -80,11 +80,14 @@ typedef enum JNX_HTTP_STATE {
  * @params out_len is the length of outward reply
  * @return message length sent in bytes 
  */
-size_t jnx_http_request(JNX_HTTP_TYPE type, const jnx_char *hostname, const jnx_char *page, jnx_char *args, jnx_uint8 **out_reply,jnx_size *out_len);
+size_t jnx_http_request(JNX_HTTP_TYPE type, const jnx_char *hostname, 
+    const jnx_char *page, jnx_char *args, jnx_uint8 **out_reply,jnx_size *out_len);
 
-JNX_HTTP_TYPE jnx_http_request_post(const jnx_char *hostname, const jnx_char *page, jnx_char *args,jnx_uint8 **out_reply, jnx_size *out_len);
+JNX_HTTP_TYPE jnx_http_request_post(const jnx_char *hostname, 
+    const jnx_char *page, jnx_char *args,jnx_uint8 **out_reply, jnx_size *out_len);
 
-JNX_HTTP_TYPE jnx_http_request_get(const jnx_char *hostname, const jnx_char *page, jnx_char *args,jnx_uint8 **out_reply, jnx_size *out_len);
+JNX_HTTP_TYPE jnx_http_request_get(const jnx_char *hostname,
+    const jnx_char *page, jnx_char *args,jnx_uint8 **out_reply, jnx_size *out_len);
 
 #ifdef __cplusplus
 }
